deepsolar-analysis (.Rmd)
The main task of the project is to predict the target variable in the given data set using supervised classification methods given the set of predictive variables. The complete data analysis is carried out with those supervised methods explained in the course, finally, the best model is chosen with the highest accuracy.